+++ src/OdfElementName.ts
@@ -1,12 +1,10 @@
 export enum OdfElementName {
-  DrawFrame = "draw:frame",
-  DrawImage = "draw:image",
-
   OfficeAutomaticStyles = "office:automatic-styles",
   OfficeBinaryData = "office:binary-data",
   OfficeBody = "office:body",
   OfficeDocument = "office:document",
   OfficeFontFaceDeclarations = "office:font-face-decls",
+  OfficeMeta = "office:meta",
   OfficeText = "office:text",
 
   StyleFontFace = "style:font-face",
@@ -15,13 +13,4 @@ export enum OdfElementName {
   StyleTabStop = "style:tab-stop",
   StyleTabStops = "style:tab-stops",
   StyleTextProperties = "style:text-properties",
-
-  TextHeading = "text:h",
-  TextHyperlink = "text:a",
-  TextLineBreak = "text:line-break",
-  TextList = "text:list",
-  TextListItem = "text:list-item",
-  TextParagraph = "text:p",
-  TextSpace = "text:s",
-  TextTabulation = "text:tab",
 }