+++ src/draw/DrawElementName.ts
@@ -1,0 +1,4 @@
+export enum DrawElementName {
+  DrawFrame = "draw:frame",
+  DrawImage = "draw:image",
+}